requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# LLM & Embeddings
openai==1.3.0

# Vector Database
pinecone-client==2.2.4

# Database
supabase==2.0.0

# HTTP Client
httpx

# Utilities
python-dotenv==1.0.0
structlog==23.2.0
tenacity==8.2.3
jinja2==3.1.2

# Development
pytest==7.4.3
pytest-asyncio==0.21.1